% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jugglr-package.R
\docType{package}
\name{jugglr-package}
\alias{jugglr}
\alias{jugglr-package}
\title{jugglr: Juggling Pattern Validation and Visualisation}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Validate and visualise juggling patterns expressed in siteswap notation. Supports vanilla, synchronous, multiplex, synchronous multiplex, and passing siteswap, with tools to check pattern validity and to retrieve raw throw data for custom visualisations. Patterns can be plotted as timeline arc diagrams or ladder diagrams, and animated via the 'JugglingLab' animation server (\url{https://jugglinglab.org/html/animinfo.html}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EllaKaye/jugglr}
  \item \url{https://ellakaye.github.io/jugglr/}
  \item Report bugs at \url{https://github.com/EllaKaye/jugglr/issues}
}

}
\author{
\strong{Maintainer}: Ella Kaye \email{hello@ellakaye.co.uk} [copyright holder]

Authors:
\itemize{
  \item Ella Kaye \email{hello@ellakaye.co.uk} [copyright holder]
}

}
\keyword{internal}
