% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{animate}
\alias{animate}
\alias{animate_markdown}
\title{Embed juggling animation in R Markdown or Quarto}
\usage{
animate(pattern, colors = NULL, ..., path = NULL)

animate_markdown(pattern, path, colors = NULL, ...)
}
\arguments{
\item{pattern}{Siteswap pattern}

\item{colors}{Optional vector of colors}

\item{...}{Additional arguments to JugglingLab GIT server}

\item{path}{Path to save GIF. If \code{NULL} (default) view in viewer}
}
\value{
knitr graphics object for embedding
}
\description{
Embed juggling animation in R Markdown or Quarto
}
