% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jugglr-package.R
\docType{package}
\name{jugglr-package}
\alias{jugglr}
\alias{jugglr-package}
\title{jugglr: Juggling Pattern Validation and Visualisation}
\description{
Validate and visualise juggling patterns expressed in siteswap notation. Supports vanilla, synchronous, multiplex, synchronous multiplex, and passing siteswap, with tools to check pattern validity, produce timeline arc diagrams and ladder diagrams, retrieve raw throw data for custom visualisations, and animate patterns via the JugglingLab GIF server.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EllaKaye/jugglr}
  \item \url{https://ellakaye.github.io/jugglr/}
  \item Report bugs at \url{https://github.com/EllaKaye/jugglr/issues}
}

}
\author{
\strong{Maintainer}: Ella Kaye \email{ella.kaye@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Ella Kaye \email{ella.kaye@gmail.com} [copyright holder]
}

}
\keyword{internal}
