% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jugglr-package.R
\docType{package}
\name{jugglr-package}
\alias{jugglr}
\alias{jugglr-package}
\title{jugglr: Juggling Pattern Validation and Visualisation}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Ella Kaye \email{ella.kaye@gmail.com} [copyright holder]

}
\keyword{internal}
