% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{timeline}
\alias{timeline}
\title{Timeline}
\usage{
timeline(siteswap, ...)
}
\arguments{
\item{siteswap}{A siteswap object}

\item{...}{Additional arguments passed to methods}
}
\description{
Plot the timeline
}
